+++ accounting/authentication/jwt.py
@@ -2,7 +2,7 @@ from fastapi.security import OAuth2PasswordBearer
 from fastapi import Depends
 from jose import JWTError, jwt
 from datetime import timedelta, datetime
-from users import User
+from accounting.users import User
 from typing import Type
 from utils.exceptions import UnauthorizedException
 SECRET_KEY = "09d25e094faa6ca2556c818166b7a9563b93f7099f6f0f4caa6cf63b88e8d3e7"
