+++ accounting/groups/__init__.py
@@ -1,0 +1,2 @@
+from .models import Group
+from .routing import group_router
