+++ accounting/roles/__init__.py
@@ -1,0 +1,2 @@
+from .models import Role
+from .routing import role_router
