+++ accounting/users/__init__.py
@@ -1,2 +1,2 @@
-from .models import User
+from .models import User, T_U
 from .routing import user_router
